requirements.txt
pandas
numpy
requests
python-dotenv
pydantic
python-dateutil
pyarrow
geopandas
osmnx
